Creation du service loadFile et injection du service dans les component landing et projets + configuration des providers
