+++ src/app/app.module.ts
@@ -15,6 +15,7 @@ import { CompetenceComponent } from './competence/competence.component';
 import { CompetencesListComponent } from './competences-list/competences-list.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { NotFoundComponent } from './not-found/not-found.component';
+import { LoadfileserviceComponent } from './loadfileservice/loadfileservice.component';
 
 register();
 
@@ -32,13 +33,14 @@ register();
     CompetenceComponent,
     CompetencesListComponent,
     NotFoundComponent,
-  ],
+    LoadfileserviceComponent
+    ],
   imports: [
     BrowserModule,
     AppRoutingModule,
     BrowserAnimationsModule
   ],
-  providers: [],
+  providers: [LoadfileserviceComponent],
   bootstrap: [AppComponent]
 })
 export class AppModule {
