+++ src/app/loadfileservice/loadfileservice.component.html
@@ -1,0 +1,1 @@
+<p>loadfileservice works!</p>
